Maher v1.1-seguridad colocada
se coloco la seguridad en todos los controllers

+++ src/prueba/pruebaBundle/Controller/GastosoperativosController.php
@@ -22,15 +22,23 @@ class GastosoperativosController extends Controller
      * @Route("/", name="gastosoperativos_index")
      * @Method("GET")
      */
-    public function indexAction()
+    public function indexAction(Request $request)
     {
-        $em = $this->getDoctrine()->getManager();
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $em = $this->getDoctrine()->getManager();
 
-        $gastosoperativos = $em->getRepository('pruebaBundle:Gastosoperativos')->findAll();
+            $gastosoperativos = $em->getRepository('pruebaBundle:Gastosoperativos')->findAll();
 
-        return $this->render('gastosoperativos/index.html.twig', array(
-            'gastosoperativos' => $gastosoperativos,
-        ));
+            return $this->render('gastosoperativos/index.html.twig', array(
+                        'gastosoperativos' => $gastosoperativos,
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
+        }
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -41,22 +49,30 @@ class GastosoperativosController extends Controller
      */
     public function newAction(Request $request)
     {
-        $gastosoperativo = new Gastosoperativos();
-        $form = $this->createForm('prueba\pruebaBundle\Form\GastosoperativosType', $gastosoperativo);
-        $form->handleRequest($request);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $gastosoperativo = new Gastosoperativos();
+            $form = $this->createForm('prueba\pruebaBundle\Form\GastosoperativosType', $gastosoperativo);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($gastosoperativo);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($gastosoperativo);
+                $em->flush();
 
-            return $this->redirectToRoute('gastosoperativos_show', array('id' => $gastosoperativo->getId()));
+                return $this->redirectToRoute('gastosoperativos_show', array('id' => $gastosoperativo->getId()));
+            }
+
+            return $this->render('gastosoperativos/new.html.twig', array(
+                        'gastosoperativo' => $gastosoperativo,
+                        'form' => $form->createView(),
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
         }
-
-        return $this->render('gastosoperativos/new.html.twig', array(
-            'gastosoperativo' => $gastosoperativo,
-            'form' => $form->createView(),
-        ));
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -67,12 +83,20 @@ class GastosoperativosController extends Controller
      */
     public function showAction(Gastosoperativos $gastosoperativo)
     {
-        $deleteForm = $this->createDeleteForm($gastosoperativo);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $deleteForm = $this->createDeleteForm($gastosoperativo);
 
-        return $this->render('gastosoperativos/show.html.twig', array(
-            'gastosoperativo' => $gastosoperativo,
-            'delete_form' => $deleteForm->createView(),
-        ));
+            return $this->render('gastosoperativos/show.html.twig', array(
+                        'gastosoperativo' => $gastosoperativo,
+                        'delete_form' => $deleteForm->createView(),
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
+        }
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -83,23 +107,31 @@ class GastosoperativosController extends Controller
      */
     public function editAction(Request $request, Gastosoperativos $gastosoperativo)
     {
-        $deleteForm = $this->createDeleteForm($gastosoperativo);
-        $editForm = $this->createForm('prueba\pruebaBundle\Form\GastosoperativosType', $gastosoperativo);
-        $editForm->handleRequest($request);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $deleteForm = $this->createDeleteForm($gastosoperativo);
+            $editForm = $this->createForm('prueba\pruebaBundle\Form\GastosoperativosType', $gastosoperativo);
+            $editForm->handleRequest($request);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($gastosoperativo);
-            $em->flush();
+            if ($editForm->isSubmitted() && $editForm->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($gastosoperativo);
+                $em->flush();
 
-            return $this->redirectToRoute('gastosoperativos_edit', array('id' => $gastosoperativo->getId()));
+                return $this->redirectToRoute('gastosoperativos_edit', array('id' => $gastosoperativo->getId()));
+            }
+
+            return $this->render('gastosoperativos/edit.html.twig', array(
+                        'gastosoperativo' => $gastosoperativo,
+                        'edit_form' => $editForm->createView(),
+                        'delete_form' => $deleteForm->createView(),
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
         }
-
-        return $this->render('gastosoperativos/edit.html.twig', array(
-            'gastosoperativo' => $gastosoperativo,
-            'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        ));
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -110,16 +142,24 @@ class GastosoperativosController extends Controller
      */
     public function deleteAction(Request $request, Gastosoperativos $gastosoperativo)
     {
-        $form = $this->createDeleteForm($gastosoperativo);
-        $form->handleRequest($request);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $form = $this->createDeleteForm($gastosoperativo);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($gastosoperativo);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($gastosoperativo);
+                $em->flush();
+            }
+
+            return $this->redirectToRoute('gastosoperativos_index');
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
         }
-
-        return $this->redirectToRoute('gastosoperativos_index');
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -131,10 +171,18 @@ class GastosoperativosController extends Controller
      */
     private function createDeleteForm(Gastosoperativos $gastosoperativo)
     {
-        return $this->createFormBuilder()
-            ->setAction($this->generateUrl('gastosoperativos_delete', array('id' => $gastosoperativo->getId())))
-            ->setMethod('DELETE')
-            ->getForm()
-        ;
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            return $this->createFormBuilder()
+                            ->setAction($this->generateUrl('gastosoperativos_delete', array('id' => $gastosoperativo->getId())))
+                            ->setMethod('DELETE')
+                            ->getForm()
+            ;
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
+        }
+        return $this->redirect($this->generateUrl('inicio'));
     }
 }
