+++ src/prueba/pruebaBundle/Controller/PagochequeController.php
@@ -22,15 +22,23 @@ class PagochequeController extends Controller
      * @Route("/", name="pagocheque_index")
      * @Method("GET")
      */
-    public function indexAction()
+    public function indexAction(Request $request)
     {
-        $em = $this->getDoctrine()->getManager();
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $em = $this->getDoctrine()->getManager();
 
-        $pagocheques = $em->getRepository('pruebaBundle:Pagocheque')->findAll();
+            $pagocheques = $em->getRepository('pruebaBundle:Pagocheque')->findAll();
 
-        return $this->render('pagocheque/index.html.twig', array(
-            'pagocheques' => $pagocheques,
-        ));
+            return $this->render('pagocheque/index.html.twig', array(
+                        'pagocheques' => $pagocheques,
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
+        }
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -41,22 +49,30 @@ class PagochequeController extends Controller
      */
     public function newAction(Request $request)
     {
-        $pagocheque = new Pagocheque();
-        $form = $this->createForm('prueba\pruebaBundle\Form\PagochequeType', $pagocheque);
-        $form->handleRequest($request);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $pagocheque = new Pagocheque();
+            $form = $this->createForm('prueba\pruebaBundle\Form\PagochequeType', $pagocheque);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($pagocheque);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($pagocheque);
+                $em->flush();
 
-            return $this->redirectToRoute('pagocheque_show', array('id' => $pagocheque->getId()));
+                return $this->redirectToRoute('pagocheque_show', array('id' => $pagocheque->getId()));
+            }
+
+            return $this->render('pagocheque/new.html.twig', array(
+                        'pagocheque' => $pagocheque,
+                        'form' => $form->createView(),
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
         }
-
-        return $this->render('pagocheque/new.html.twig', array(
-            'pagocheque' => $pagocheque,
-            'form' => $form->createView(),
-        ));
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -67,12 +83,20 @@ class PagochequeController extends Controller
      */
     public function showAction(Pagocheque $pagocheque)
     {
-        $deleteForm = $this->createDeleteForm($pagocheque);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $deleteForm = $this->createDeleteForm($pagocheque);
 
-        return $this->render('pagocheque/show.html.twig', array(
-            'pagocheque' => $pagocheque,
-            'delete_form' => $deleteForm->createView(),
-        ));
+            return $this->render('pagocheque/show.html.twig', array(
+                        'pagocheque' => $pagocheque,
+                        'delete_form' => $deleteForm->createView(),
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
+        }
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -83,23 +107,31 @@ class PagochequeController extends Controller
      */
     public function editAction(Request $request, Pagocheque $pagocheque)
     {
-        $deleteForm = $this->createDeleteForm($pagocheque);
-        $editForm = $this->createForm('prueba\pruebaBundle\Form\PagochequeType', $pagocheque);
-        $editForm->handleRequest($request);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $deleteForm = $this->createDeleteForm($pagocheque);
+            $editForm = $this->createForm('prueba\pruebaBundle\Form\PagochequeType', $pagocheque);
+            $editForm->handleRequest($request);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($pagocheque);
-            $em->flush();
+            if ($editForm->isSubmitted() && $editForm->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($pagocheque);
+                $em->flush();
 
-            return $this->redirectToRoute('pagocheque_edit', array('id' => $pagocheque->getId()));
+                return $this->redirectToRoute('pagocheque_edit', array('id' => $pagocheque->getId()));
+            }
+
+            return $this->render('pagocheque/edit.html.twig', array(
+                        'pagocheque' => $pagocheque,
+                        'edit_form' => $editForm->createView(),
+                        'delete_form' => $deleteForm->createView(),
+            ));
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
         }
-
-        return $this->render('pagocheque/edit.html.twig', array(
-            'pagocheque' => $pagocheque,
-            'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        ));
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -110,16 +142,24 @@ class PagochequeController extends Controller
      */
     public function deleteAction(Request $request, Pagocheque $pagocheque)
     {
-        $form = $this->createDeleteForm($pagocheque);
-        $form->handleRequest($request);
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            $form = $this->createDeleteForm($pagocheque);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($pagocheque);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($pagocheque);
+                $em->flush();
+            }
+
+            return $this->redirectToRoute('pagocheque_index');
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
         }
-
-        return $this->redirectToRoute('pagocheque_index');
+        return $this->redirect($this->generateUrl('inicio'));
     }
 
     /**
@@ -131,10 +171,18 @@ class PagochequeController extends Controller
      */
     private function createDeleteForm(Pagocheque $pagocheque)
     {
-        return $this->createFormBuilder()
-            ->setAction($this->generateUrl('pagocheque_delete', array('id' => $pagocheque->getId())))
-            ->setMethod('DELETE')
-            ->getForm()
-        ;
+        $session = $request->getSession();
+        if ($session->get('tipousuario') == 'Administrador' or $session->get('tipousuario') == 'Empleado') {
+            return $this->createFormBuilder()
+                            ->setAction($this->generateUrl('pagocheque_delete', array('id' => $pagocheque->getId())))
+                            ->setMethod('DELETE')
+                            ->getForm()
+            ;
+        } else {
+            $this->get('session')->getFlashBag()->add(
+                    'Mensaje', "Esta intentando entrar a una zona de seguridad a la cual no tiene acceso"
+            );
+        }
+        return $this->redirect($this->generateUrl('inicio'));
     }
 }
